Fixed country list, and a typo.
I got really really OCD and spent like 1 and a half hours fixing the countries. I also fixed one of the typos, but not all of them.

Some of them can't be read clearly, so I had to pretty much figure it out from spacing and guesses. Two can't be guessed at all: INADVERTENT (Something) and the line directly below, because of one particular guy who walked in front of the camera making me rage because I couldn't even see it on the main monitor to read it there, and even if I could have read it off of a reflection some other guy in a white shirt walked in front of it the one frame even that was there causing me to swear out loud. I just just of bullshitted in INADVERTENT INCIDENT, and U.S. DOMESTIC, because I thought they were funny. I needed funny after getting mad at that.

I need to sleep now.